Migration system implementation #167
