+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/config/IMigrationHelper.java
@@ -1,0 +1,5 @@
+package io.github.wysohn.triggerreactor.core.manager.config;
+
+public interface IMigrationHelper {
+    void migrate(IConfigSource current);
+}
