+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/trigger/AbstractCommandTriggerManager.java
@@ -18,7 +18,6 @@ package io.github.wysohn.triggerreactor.core.manager.trigger;
 
 import io.github.wysohn.triggerreactor.core.bridge.ICommandSender;
 import io.github.wysohn.triggerreactor.core.main.TriggerReactor;
-import io.github.wysohn.triggerreactor.core.script.wrapper.SelfReference;
 import io.github.wysohn.triggerreactor.tools.FileUtil;
 
 import java.io.File;
@@ -202,8 +201,8 @@ public abstract class AbstractCommandTriggerManager extends AbstractTriggerManag
         }
     }
 
-    public AbstractCommandTriggerManager(TriggerReactor plugin, SelfReference ref, File tirggerFolder) {
-        super(plugin, ref, tirggerFolder);
+    public AbstractCommandTriggerManager(TriggerReactor plugin, File tirggerFolder) {
+        super(plugin, tirggerFolder);
     }
 
     public static class CommandTrigger extends Trigger {
